refactor: Split classes into separate files per StyleCop rules
- Extract CallStatus enum to SharedModels.Dtos.Contacts.CallStatus.cs
- Extract CallLogEntry record to SharedModels.Dtos.Contacts.CallLogEntry.cs
- Extract CallLog class to WebApi.Data.Models.CallLog.cs
- Remove duplicate CallStatus enum from WebApi.Data.Models
- Update all references to use shared CallStatus from SharedModels

This addresses SA1402 (File may only contain a single type) warnings
and eliminates duplicate enum definitions between SharedModels and WebApi.
