+++ src/SharedModels/Dtos/Contacts/CallLogEntry.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace SharedModels.Dtos.Contacts
+{
+    public record CallLogEntry(
+        DateTime CallDate,
+        string Notes,
+        CallStatus Status,
+        DateTime? NextCallDate = null
+    );
+}
