+++ src/SharedModels/Dtos/Contacts/CallStatus.cs
@@ -1,0 +1,12 @@
+namespace SharedModels.Dtos.Contacts
+{
+    public enum CallStatus
+    {
+        New,
+        Scheduled,
+        Reached,
+        NotReached,
+        Completed,
+        Postponed
+    }
+}
