+++ src/SharedModels/Dtos/Contacts/ContactDto.cs
@@ -1,8 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using SharedModels.Dtos.Common;
 
 namespace SharedModels.Dtos.Contacts
@@ -19,21 +16,4 @@ namespace SharedModels.Dtos.Contacts
         DateTime? LastCallDate = null,
         List<CallLogEntry>? CallHistory = null
     ) : BaseDto;
-
-    public enum CallStatus
-    {
-        New,
-        Scheduled,
-        Reached,
-        NotReached,
-        Completed,
-        Postponed
-    }
-
-    public record CallLogEntry(
-        DateTime CallDate,
-        string Notes,
-        CallStatus Status,
-        DateTime? NextCallDate = null
-    );
 }
